requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
python-dotenv==1.0.1
tzdata==2024.1
pytest==8.3.2